requirements.txt
GitPython>=3.1
unidiff>=0.6
google-genai>=0.6
typer
python-dotenv
markdown
weasyprint
uvicorn
fastapi